fix bug

+++ src/DataTypes.java
@@ -1,8 +1,7 @@
 import java.util.List;
 
 public class DataTypes {
-    // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
-    public static long sum(List<Integer> numbers) {
+    public static int sum(List<Integer> numbers) {
 
         int s = 0;
         // below is a "foreach" loop which iterates through numbers
